@@ -2,24 +2,29 @@ package com.shultzlab.weighttrackerapi.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-@EnableWebSecurity
-@Configuration
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+/*
+The objective of the config is to enable CORS
+ */
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.headers().xssProtection();
+@Configuration
+public class WebSecurityConfig {
+
+    @Bean
+    public DefaultSecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
                 .cors().configurationSource(corsConfigurationSource())
                 .and()
-                .csrf().disable();
+                .csrf().disable()
+                .headers().xssProtection();
+
+        return http.build();
     }
 
     @Bean
@@ -32,4 +37,5 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
         source.registerCorsConfiguration("/**", configuration);
         return source;
     }
+
 }
