@@ -3,11 +3,11 @@ package com.shultzlab.weighttrackerapi;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-@SpringBootApplication
+
+@SpringBootApplication()
 public class WeightTrackerApiApplication {
 
     public static void main(String[] args) {
